WIP Fix image properties
