+++ src/plugins/ai-assistant/ai-assistant.ts
@@ -20,7 +20,7 @@ import { Plugin } from 'jodit/core/plugin/plugin';
 import './config';
 
 import { UiAiAssistant } from './ui/ui-ai-assistant';
-import * as langs from './langs';
+import * as lang from './langs';
 
 /**
  * The plugin inserts content generated by AI into the editor.
@@ -71,11 +71,11 @@ export class aiAssistant extends Plugin {
 
 	constructor(jodit: IJodit) {
 		super(jodit);
-		extendLang(langs);
+		extendLang(lang);
 	}
 
 	/** @override */
-	override afterInit(jodit: IJodit): void {}
+	override afterInit(): void {}
 
 	@watch(':generateAiAssistantForm.ai-assistant')
 	protected onGenerateAiAssistantForm(prompt: string): void {
